+++ sklearn/neural_network/__init__.py
@@ -5,10 +5,10 @@ networks.
 
 # License: BSD 3 clause
 
-from .rbm import BernoulliRBM
+from ._rbm import BernoulliRBM
 
-from .multilayer_perceptron import MLPClassifier
-from .multilayer_perceptron import MLPRegressor
+from ._multilayer_perceptron import MLPClassifier
+from ._multilayer_perceptron import MLPRegressor
 
 __all__ = ["BernoulliRBM",
            "MLPClassifier",
